refactor(MagicLink): Split in files
